+++ lib/kiq/running_job.ex
@@ -1,7 +1,7 @@
 defmodule Kiq.RunningJob do
   @moduledoc false
 
-  alias Kiq.{Job, Timestamp}
+  alias Kiq.{Encoder, Job, Timestamp}
 
   @type t :: %__MODULE__{key: binary(), encoded: binary()}
 
@@ -11,7 +11,7 @@ defmodule Kiq.RunningJob do
   def new(%Job{pid: pid, queue: queue} = job) do
     details = %{queue: queue, payload: Job.to_map(job), run_at: Timestamp.unix_now()}
 
-    %__MODULE__{key: format_key(pid), encoded: Jason.encode!(details)}
+    %__MODULE__{key: format_key(pid), encoded: Encoder.encode(details)}
   end
 
   defp format_key(pid) when is_pid(pid) or is_reference(pid) do